Fix: Flatten directory structure
